Rename message to translations
