+++ suripu-core/src/main/java/com/hello/suripu/core/models/Events/SleepEvent.java
@@ -2,7 +2,7 @@ package com.hello.suripu.core.models.Events;
 
 import com.hello.suripu.core.models.Event;
 import com.hello.suripu.core.models.SleepSegment;
-import com.hello.suripu.core.message.English;
+import com.hello.suripu.core.translations.English;
 
 /**
  * Created by pangwu on 11/24/14.
